Avoid using TimeSub in Spark3.1+ and move First/Last out of shims
